requirements-tests.txt
coverage
cryptography~=35.0.0
flask~=2.0.2
mock~=4.0.3
parameterized~=0.8.1
pylint
pytest
requests~=2.26.0
respx
sap_py_jwt
tox
